+++ Samples/luainvaders/src/Mothership.h
@@ -45,7 +45,7 @@ public:
 
 private:
 	// Time of the last update
-	float update_frame_start;
+	double update_frame_start;
 
 	// Direction mothership is flying in
 	float direction;
