+++ Source/Controls/Lua/ElementFormControlInput.cpp
@@ -104,7 +104,7 @@ int ElementFormControlInputSetAttrmaxlength(lua_State* L)
 {
     ElementFormControlInput* obj = LuaType<ElementFormControlInput>::check(L,1);
     LUACHECKOBJ(obj);
-    int maxlength = luaL_checkinteger(L,2);
+    int maxlength = (int)luaL_checkinteger(L,2);
     obj->SetAttribute("maxlength",maxlength);
     return 0;
 }
@@ -113,7 +113,7 @@ int ElementFormControlInputSetAttrsize(lua_State* L)
 {
     ElementFormControlInput* obj = LuaType<ElementFormControlInput>::check(L,1);
     LUACHECKOBJ(obj);
-    int size = luaL_checkinteger(L,2);
+    int size = (int)luaL_checkinteger(L,2);
     obj->SetAttribute("size",size);
     return 0;
 }
@@ -122,7 +122,7 @@ int ElementFormControlInputSetAttrmax(lua_State* L)
 {
     ElementFormControlInput* obj = LuaType<ElementFormControlInput>::check(L,1);
     LUACHECKOBJ(obj);
-    int max = luaL_checkinteger(L,2);
+    int max = (int)luaL_checkinteger(L,2);
     obj->SetAttribute("max",max);
     return 0;
 }
@@ -131,7 +131,7 @@ int ElementFormControlInputSetAttrmin(lua_State* L)
 {
     ElementFormControlInput* obj = LuaType<ElementFormControlInput>::check(L,1);
     LUACHECKOBJ(obj);
-    int min = luaL_checkinteger(L,2);
+    int min = (int)luaL_checkinteger(L,2);
     obj->SetAttribute("min",min);
     return 0;
 }
@@ -140,7 +140,7 @@ int ElementFormControlInputSetAttrstep(lua_State* L)
 {
     ElementFormControlInput* obj = LuaType<ElementFormControlInput>::check(L,1);
     LUACHECKOBJ(obj);
-    int step = luaL_checkinteger(L,2);
+    int step = (int)luaL_checkinteger(L,2);
     obj->SetAttribute("step",step);
     return 0;
 }
