+++ Source/Controls/Lua/ElementFormControlSelect.cpp
@@ -47,7 +47,7 @@ int ElementFormControlSelectAdd(lua_State* L, ElementFormControlSelect* obj)
     const char* value = luaL_checkstring(L,2);
     int before = -1; //default
     if(lua_gettop(L) >= 3)
-        before = luaL_checkinteger(L,3);
+        before = (int)luaL_checkinteger(L,3);
 
     int index = obj->Add(rml,value,before);
     lua_pushinteger(L,index);
@@ -56,7 +56,7 @@ int ElementFormControlSelectAdd(lua_State* L, ElementFormControlSelect* obj)
 
 int ElementFormControlSelectRemove(lua_State* L, ElementFormControlSelect* obj)
 {
-    int index = luaL_checkinteger(L,1);
+    int index = (int)luaL_checkinteger(L,1);
     obj->Remove(index);
     return 0;
 }
@@ -87,7 +87,7 @@ int ElementFormControlSelectSetAttrselection(lua_State* L)
 {
     ElementFormControlSelect* obj = LuaType<ElementFormControlSelect>::check(L,1);
     LUACHECKOBJ(obj);
-    int selection = luaL_checkinteger(L,2);
+    int selection = (int)luaL_checkinteger(L,2);
     obj->SetSelection(selection);
     return 0;
 }
