+++ Source/Controls/Lua/ElementFormControlTextArea.cpp
@@ -75,7 +75,7 @@ int ElementFormControlTextAreaSetAttrcols(lua_State* L)
 {
     ElementFormControlTextArea* obj = LuaType<ElementFormControlTextArea>::check(L,1);
     LUACHECKOBJ(obj);
-    int cols = luaL_checkinteger(L,2);
+    int cols = (int)luaL_checkinteger(L,2);
     obj->SetNumColumns(cols);
     return 0;
 }
@@ -84,7 +84,7 @@ int ElementFormControlTextAreaSetAttrmaxlength(lua_State* L)
 {
     ElementFormControlTextArea* obj = LuaType<ElementFormControlTextArea>::check(L,1);
     LUACHECKOBJ(obj);
-    int ml = luaL_checkinteger(L,2);
+    int ml = (int)luaL_checkinteger(L,2);
     obj->SetMaxLength(ml);
     return 0;
 }
@@ -93,7 +93,7 @@ int ElementFormControlTextAreaSetAttrrows(lua_State* L)
 {
     ElementFormControlTextArea* obj = LuaType<ElementFormControlTextArea>::check(L,1);
     LUACHECKOBJ(obj);
-    int rows = luaL_checkinteger(L,2);
+    int rows = (int)luaL_checkinteger(L,2);
     obj->SetNumRows(rows);
     return 0;
 }
