+++ Source/Core/Lua/Vector2i.cpp
@@ -59,8 +59,8 @@ template<> void ExtraInit<Vector2i>(lua_State* L, int metatable_index)
 
 int Vector2inew(lua_State* L)
 {
-    int x = luaL_checkinteger(L,1);
-    int y = luaL_checkinteger(L,2);
+    int x = (int)luaL_checkinteger(L,1);
+    int y = (int)luaL_checkinteger(L,2);
 
     Vector2i* vect = new Vector2i(x,y);
 
@@ -72,7 +72,7 @@ int Vector2i__mul(lua_State* L)
 {
     Vector2i* lhs = LuaType<Vector2i>::check(L,1);
     LUACHECKOBJ(lhs);
-    int rhs = luaL_checkinteger(L,2);
+    int rhs = (int)luaL_checkinteger(L,2);
 
     Vector2i* res = new Vector2i(0,0);
     (*res) = (*lhs) * rhs;
@@ -85,7 +85,7 @@ int Vector2i__div(lua_State* L)
 {
     Vector2i* lhs = LuaType<Vector2i>::check(L,1);
     LUACHECKOBJ(lhs);
-    int rhs = luaL_checkinteger(L,2);
+    int rhs = (int)luaL_checkinteger(L,2);
 
     Vector2i* res = new Vector2i(0,0);
     (*res) = (*lhs) / rhs;
@@ -164,7 +164,7 @@ int Vector2iSetAttrx(lua_State*L)
 {
     Vector2i* self = LuaType<Vector2i>::check(L,1);
     LUACHECKOBJ(self);
-    int value = luaL_checkinteger(L,2);
+    int value = (int)luaL_checkinteger(L,2);
 
     self->x = value;
     return 0;
@@ -174,7 +174,7 @@ int Vector2iSetAttry(lua_State*L)
 {
     Vector2i* self = LuaType<Vector2i>::check(L,1);
     LUACHECKOBJ(self);
-    int value = luaL_checkinteger(L,2);
+    int value = (int)luaL_checkinteger(L,2);
 
     self->y = value;
     return 0;
